Convert compress option to stream_handler for Cowboy2 adapter
In order to maintain backwards compatability, if the compress option is
set then the `cowboy_compress_h` handler is used. This is how
compression is configured in Cowboy 2. This is not compatible with the
users setting their own stream_handlers.

This was requested in https://github.com/VoiceLayer/plug_cowboy2/issues/3#issuecomment-331083574