Fix start function for Cowboy 2 adapter (#618)
Previously, the cowboy start_clear/start_tls functions were being used directly.
This meant that they were being started under the supervision tree for ranch and
not the supervision tree for the actual application. This approach use
`ranch_listener_sup` as the entry point instead. This matches what happens in
the Cowboy adapter.