Update Cowboy 2 adapter to use Ranch 1.6 options format (#778)
* Update Cowboy 2 adapter to use Ranch 1.6 options format

Cowboy has been updated to 2.5 in mix.exs. Cowboy 2.5 depends on Ranch
1.6.2. This version off ranch expects a map as the transport options,
with the socket options being a list inside of the `:socket_opts` key.

* Explicitly require transport_options for Cowboy 2

When using the Cowboy 2 adapter, the `transport_options` must now be explicitly
set to override the defaults. The previous `:acceptors` and `:max_connections`
options have been deprecated in favour of passing them via the
`transport_options`.

* Update cowboy2.ex