Make the project a valid mix project
All of the tests are passing. The module has been renamed to `Plug.Cowboy` and
any references to it in the test have been updated.

The files have been moved to more naturally reflect the module names, however
the `Conn` and `Handler` modules have not been renamed for compatibility
reasons.