Speed up inform test by setting cowboy timeouts
The hackney client is slow on the inform tests as it does not handle 1xx
repsonses correctly, and waits for the end of the request. When the
request times out, the response is successfully returned, causing the
test to pass.

These timeouts have been redueced to 1s to ensure the tests return in a
reasonable amount of time, instead of the cowboy defaults of 5s for the
`idle_timeout` and 60s for the `request_timeout`.